requirements-web.txt
# Web Application Requirements
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Audio Processing
sounddevice==0.4.6
numpy==1.24.3

# HTTP Requests
requests==2.31.0

# Audio File Processing
wave
tempfile
os
json
threading
queue
time
datetime
typing
